+++ src/main/java/com/misterymuscle/springapisample/dto/BoardArticleDto.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class BoardArticleDto {
 
-    private int boardIdx;
+    private long boardIdx;
     private String boardTitle;
     private String boardContent;
     private Date boardRegDate;
